raspberrypi/requirements.txt
awsiotsdk
python-dotenv
pythonping
